[search] Fix uint16 wrapping bug for party/alliance leaders
This fixes search for alliance/party IDs above 65535.
The IDs are mapped to player IDs and needed to be widened.

+++ src/search/data_loader.h
@@ -89,7 +89,7 @@ public:
     uint32 GetPlayersCount(const search_req& sr);
 
     std::vector<ahHistory*>  GetAHItemHystory(uint16 ItemID, bool stack);
-    std::list<SearchEntity*> GetPartyList(uint16 PartyID, uint16 AllianceID);
+    std::list<SearchEntity*> GetPartyList(uint32 PartyID, uint32 AllianceID);
     std::list<SearchEntity*> GetLinkshellList(uint32 LinkshellID);
     std::list<SearchEntity*> GetPlayersList(search_req sr, int* count);
     std::string              GetSearchComment(uint32 playerId);
